requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose
python-dotenv
pydantic